AI/requirements.txt
pymongo
certifi
openai
sentence-transformers
python-dotenv
flask
flask-cors
pymupdf
requests

# YOLO Preprocessing Dependencies
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
easyocr>=1.7.0
shapely>=2.0.0

# LoRA Fine-tuning and Inference Dependencies
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.28.0
datasets>=2.18.0
pillow>=10.0.0
